sdk/python/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "ail-sdk"
version = "0.1.0"
description = "Python SDK for 22B Labs Agent Identity Layer"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
keywords = ["agent-identity", "22blabs", "ail", "credential", "jwt"]
dependencies = [
    "cryptography>=42.0.0",
    "PyJWT>=2.8.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0"]

[tool.setuptools.packages.find]
where = ["."]
include = ["ail_sdk*"]
